Update optimization method (#106)
* Set DE, instead of GA, as a default optimizer

* Rename options

* Set workers to 1

* Update external_optimizer

* Update tests

* Import DE here

* Update ExternalOptimizer

* Fix path

* Update docs

* Fix popsize

* Fix maxiter

* Update external optimizer

* Use _get_n_iter

* Remove test for ExternalOptimizer

* Rename Optimizer

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update trace_obj

* Remove unused import

* Marge ExecModel into ModelObject

* Add create_model

* Fix load_param

* Allow to set workers greater than 1

* Move GA code to _legacy

* Fix docs

* Save at legacy.functions

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci
