+++ biomass/core.py
@@ -1,10 +1,9 @@
 """BioMASS core functions"""
-import multiprocessing
 import os
 from dataclasses import dataclass
 from importlib import import_module
 from pathlib import Path
-from typing import Any, Callable, Dict, Iterable, Optional, Union
+from typing import Any, Callable, Dict, Optional, Union
 
 try:  # python 3.8+
     from typing import Literal
@@ -12,13 +11,14 @@ except ImportError:
     from typing_extensions import Literal
 
 import numpy as np
+from scipy.optimize import differential_evolution
 
 from .analysis import InitialConditionSensitivity, ParameterSensitivity, ReactionSensitivity
 from .dynamics import SignalingSystems
-from .estimation import GeneticAlgorithmContinue, GeneticAlgorithmInit
+from .estimation import Optimizer
 from .exec_model import ModelObject
 
-__all__ = ["Model", "optimize", "optimize_continue", "run_simulation", "run_analysis"]
+__all__ = ["Model", "create_model", "optimize", "run_simulation", "run_analysis"]
 
 
 class BiomassIndexError(Exception):
@@ -189,34 +189,39 @@ class Model(object):
         return model
 
 
-def _check_optional_arguments(
-    x_id: Union[int, Iterable[int]],
-    options: Optional[dict],
-) -> None:
-    if options is None:
-        pass
-    elif isinstance(options, dict):
-        if options["local_search_method"].lower() not in ["mutation", "powell", "de"]:
-            raise ValueError(
-                f"'{options['local_search_method']}': "
-                "Invalid local_search_method. Should be one of ['mutation', 'Powell', 'DE']"
-            )
-        elif (
-            not isinstance(x_id, int)
-            and options["local_search_method"].lower() == "de"
-            and options["workers"] != 1
-        ):
-            raise AssertionError(
-                "daemonic processes are not allowed to have children. Set options['workers'] to 1."
-            )
-    else:
-        raise TypeError("options must be dict or None.")
+def create_model(pkg_name: str, show_info: bool = False) -> ModelObject:
+    """
+    Create a BioMASS model.
+
+    Parameters
+    ----------
+    pkg_name : str
+        Path (dot-sepalated) to a biomass model directory.
+        Use ``__package__``.
+    show_info : bool (default: :obj:`False`)
+        Set to :obj:`True` to print the information related to model size.
+
+    Returns
+    -------
+    model : :class:`biomass.exec_model.ModelObject`
+        The BioMASS model object.
+
+    Examples
+    --------
+    >>> from biomass import create_model
+    >>> import your_model
+    >>> model = create_model(your_model.__package__)
+    """
+    model = Model(pkg_name).create(show_info)
+    return model
 
 
 def optimize(
     model: ModelObject,
-    x_id: Union[int, Iterable[int]],
-    options: Optional[dict] = None,
+    x_id: int,
+    *,
+    disp_here: bool = False,
+    optimizer_options: Optional[dict] = None,
 ) -> None:
     """
     Estimate model parameters from experimental data.
@@ -226,186 +231,58 @@ def optimize(
     model : ModelObject
         Model for parameter estimation.
 
-    x_id : int or Iterable[int]
-        Index (indices) of parameter set to estimate.
+    x_id : int
+        Index of parameter set to estimate.
 
-    options : dict, optional
-        * popsize : int (default: 5)
-            A multiplier for setting the total population size.
-            The population has ``popsize`` * len(`search_param`) individuals.
+    disp_here: bool (default: False)
+        Whether to show the evaluated *objective* at every iteration.
 
-        * max_generation : int (default: 10000)
-            Stop optimization if Generation > ``max_generation``.
-
-        * initial_threshold : float (default: 1e12)
-            Threshold on objective function value used to generate initial population.
-            Default value is 1e12 (numerically solvable).
-
-        * allowable_error : float (default: 0.0)
-            Stop optimization if Best Fitness <= ``allowable_error``.
-
-        * local_search_method : str (default: 'mutation')
-            Method used in local search. Should be one of:
-
-            * 'mutation' : NDM/MGG
-
-            * 'Powell' : Modified Powell method
-
-            * 'DE' : Differential Evolution (strategy: `best2bin`)
-
-        * n_children : int (default: 1000)
-            (``local_search_method`` == 'mutation') The number of children generated in NDM/MGG.
-
-        * maxiter : int (default: 10)
-            (``local_search_method`` in ['Powell', 'DE']) The maximum number of iterations
-            over which the entire population is evolved.
-
-        * workers : int (default: -1 if isinstance(x_id, `int`) else 1)
-            (``local_search_method`` == 'DE') The population is subdivided into workers sections and
-            evaluated in parallel (uses multiprocessing.Pool). Supply -1 to use
-            all available CPU cores. Set workers to 1 when searching multiple
-            parameter sets simultaneously.
-
-        * overwrite : bool (default: :obj:`False`)
-            If :obj:`True`, the out/n folder will be overwritten.
+    optimizer_options : dict, optional
+        Keyword arguments to pass to ``scipy.optimize.differential_evolution``.
+        For details, please refer to https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.differential_evolution.html.
 
     Examples
     --------
     >>> from biomass.models import Nakakuki_Cell_2010
     >>> from biomass import Model, optimize
     >>> model = Model(Nakakuki_Cell_2010.__package__).create()
-    >>> optimize(
-    ...     model, x_id=range(1, 11),
-    ...     options={
-    ...         'max_generation': 10000,
-    ...         'allowable_error': 0.5
-    ...     }
-    ... )
+    >>> optimize(model, x_id=1)
+
+    Notes
+    -----
+
+    * Set simulation conditions and the corresponding experimental data in ``observable.py``
+    * Define an objective function to be minimized (:func:`objective`) in ``fitness.py``
+    * Set lower/upper bounds of parameters to be estimated in ``set_search_param.py``
 
     """
-    if options is None:
-        options = {}
-    options.setdefault("popsize", 5)
-    options.setdefault("max_generation", 10000)
-    options.setdefault("initial_threshold", 1e12)
-    options.setdefault("allowable_error", 0.0)
-    options.setdefault("local_search_method", "mutation")
-    options.setdefault("n_children", 1000)
-    options.setdefault("maxiter", 10)
-    options.setdefault("workers", -1 if isinstance(x_id, int) else 1)
-    options.setdefault("overwrite", False)
+    if optimizer_options is None:
+        optimizer_options = {}
+    optimizer_options.setdefault("strategy", "best1bin")
+    optimizer_options.setdefault("maxiter", 50)
+    optimizer_options.setdefault("popsize", 3)
+    optimizer_options.setdefault("tol", 1e-4)
+    optimizer_options.setdefault("mutation", 0.1)
+    optimizer_options.setdefault("recombination", 0.5)
+    optimizer_options.setdefault("disp", True)
+    optimizer_options.setdefault("polish", False)
+    optimizer_options.setdefault("workers", 1)
 
-    _check_optional_arguments(x_id, options)
+    if not optimizer_options["disp"]:
+        raise ValueError(
+            "Set optimizer_options['disp'] to True. "
+            "If you don't want to see the evaluated objective function at every iteration, "
+            "set the keyword argument `disp_here` to False."
+        )
 
-    ga_init = GeneticAlgorithmInit(model, **options)
-    if isinstance(x_id, int):
-        ga_init.run(x_id)
-    elif all([isinstance(i, int) for i in x_id]):
-        n_proc = max(1, multiprocessing.cpu_count() - 1)
-        with multiprocessing.Pool(processes=n_proc) as p:
-            for _ in p.imap_unordered(ga_init.run, x_id):
-                pass
-    else:
-        raise TypeError("`x_id` type must be either int or Iterable[int].")
-
-
-def optimize_continue(
-    model: ModelObject,
-    x_id: Union[int, Iterable[int]],
-    options: Optional[dict] = None,
-) -> None:
-    """
-    Continue running optimization from where you stopped in the last parameter search.
-
-    Parameters
-    ----------
-    model : ModelObject
-        Model for parameter estimation.
-
-    x_id : int or Iterable[int]
-        Index (indices) of parameter set to estimate.
-
-    options : dict, optional
-        * popsize : int (default: 5)
-            A multiplier for setting the total population size.
-            The population has ``popsize`` * len(`search_param`) individuals.
-
-        * max_generation : int (default: 15000)
-            Stop optimization if Generation > ``max_generation``.
-
-        * initial_threshold : float (default: 1e12)
-            Threshold on objective function value used to generate initial population.
-            Default value is 1e12 (numerically solvable).
-
-        * allowable_error : float (default: 0.0)
-            Stop optimization if Best Fitness <= ``allowable_error``.
-
-        * local_search_method : str (default: 'mutation')
-            Method used in local search. Should be one of:
-
-            * 'mutation' : NDM/MGG
-
-            * 'Powell' : Modified Powell method
-
-            * 'DE' : Differential Evolution (strategy: `best2bin`)
-
-        * n_children : int (default: 1000)
-            (``local_search_method`` == 'mutation') The number of children generated in NDM/MGG.
-
-        * maxiter : int (default: 10)
-            (``local_search_method`` in ['Powell', 'DE']) The maximum number of iterations
-            over which the entire population is evolved.
-
-        * workers : int (default: -1 if isinstance(x_id, `int`) else 1)
-            (``local_search_method`` == 'DE') The population is subdivided into workers sections
-            and evaluated in parallel (uses multiprocessing.Pool). Supply -1 to use
-            all available CPU cores. Set workers to 1 when searching multiple
-            parameter sets simultaneously.
-
-        * p0_bounds : list of floats (default: [0.1, 10.0])
-            Generate initial population using best parameter values in the last parameter search.
-
-            * `lower_bound` = po_bounds[0] * `best_parameter_value`
-            * `upper_bound` = p0_bounds[1] * `best_parameter_value`
-
-    Examples
-    --------
-    >>> from biomass.models import Nakakuki_Cell_2010
-    >>> from biomass import Model, optimize_continue
-    >>> model = Model(Nakakuki_Cell_2010.__package__).create()
-    >>> optimize_continue(
-    ...     model, x_id=range(1, 11),
-    ...     options={
-    ...         'max_generation': 20000,
-    ...         'allowable_error': 0.5
-    ...     }
-    ... )
-
-    """
-    if options is None:
-        options = {}
-    options.setdefault("popsize", 5)
-    options.setdefault("max_generation", 15000)
-    options.setdefault("initial_threshold", 1e12)
-    options.setdefault("allowable_error", 0.0)
-    options.setdefault("local_search_method", "mutation")
-    options.setdefault("n_children", 1000)
-    options.setdefault("maxiter", 10)
-    options.setdefault("workers", -1 if isinstance(x_id, int) else 1)
-    options.setdefault("p0_bounds", [0.1, 10.0])
-
-    _check_optional_arguments(x_id, options)
-
-    ga_continue = GeneticAlgorithmContinue(model, **options)
-    if isinstance(x_id, int):
-        ga_continue.run(x_id)
-    elif all([isinstance(i, int) for i in x_id]):
-        n_proc = max(1, multiprocessing.cpu_count() - 1)
-        with multiprocessing.Pool(processes=n_proc) as p:
-            for _ in p.imap_unordered(ga_continue.run, x_id):
-                pass
-    else:
-        raise TypeError("`x_id` type must be either int or Iterable[int].")
+    optimizer = Optimizer(model, differential_evolution, x_id, disp_here)
+    res = optimizer.minimize(
+        model.get_obj_val,
+        [(0, 1) for _ in range(len(model.problem.bounds))],
+        **optimizer_options,
+    )
+    param_values = model.problem.gene2val(res.x)
+    optimizer.import_solution(param_values)
 
 
 def run_simulation(
