+++ biomass/dynamics/temporal_dynamics.py
@@ -7,14 +7,14 @@ import numpy as np
 from matplotlib import pyplot as plt
 from matplotlib.axes._axes import _log as matplotlib_axes_logger
 
-from ..exec_model import ExecModel, ModelObject
+from ..exec_model import ModelObject
 from ..plotting import MultipleObservables, SingleObservable
 
 matplotlib_axes_logger.setLevel("ERROR")
 
 
 @dataclass
-class TemporalDynamics(ExecModel):
+class TemporalDynamics(object):
     model: ModelObject
 
     def plot_timecourse(
