--- docs/api/estimation/external_optimizer.rst
@@ -1,5 +1,0 @@
-External Optimizer (:py:mod:`biomass.estimation.external_optimizer`)
-====================================================================
-
-.. autoclass:: biomass.estimation.ExternalOptimizer
-   :members:
