--- docs/api/estimation/ga.rst
@@ -1,9 +1,0 @@
-Genetic algorithm (:py:mod:`biomass.estimation.ga`)
-===================================================
-
-.. autoclass:: biomass.estimation.GeneticAlgorithmInit
-   :members: 
-   :private-members: _ga_v2
-
-.. autoclass:: biomass.estimation.GeneticAlgorithmContinue
-   :members:
