+++ docs/api/estimation/index.rst
@@ -4,6 +4,5 @@ Parameter estimation (:py:mod:`biomass.estimation`)
 .. toctree::
    :maxdepth: 4
 
-   ga
-   external_optimizer
+   optimizer
