+++ docs/api/estimation/optimizer.rst
@@ -1,0 +1,5 @@
+Optimizer (:py:mod:`biomass.estimation.optimizer`)
+==================================================
+
+.. autoclass:: biomass.estimation.Optimizer
+   :members:
