+++ docs/usage.rst
@@ -28,9 +28,9 @@ Name                    Content
 
 .. code-block:: python
 
-    >>> from biomass import Model
+    >>> from biomass import create_model
     >>> from biomass.models import Nakakuki_Cell_2010
-    >>> model = Model(Nakakuki_Cell_2010.__package__).create(show_info=True)
+    >>> model = create_model(Nakakuki_Cell_2010.__package__, show_info=True)
     Nakakuki_Cell_2010 information
     ------------------------------
     36 species
@@ -48,121 +48,43 @@ Using :func:`~biomass.core.optimize` function
 
 Parameters are adjusted to minimize the distance between model simulation and experimental data.
 
-.. code-block:: python
-
-    from biomass import optimize
-
-    optimize(
-        model, x_id=1, options={
-            "popsize": 3,
-            "max_generation": 100,
-            "allowable_error": 0.5,
-            "local_search_method": "DE",
-            "maxiter": 50,
-        }
-    )
-
-The temporary result will be saved in ``out/n/`` after each iteration.
-
-Progress list: ``out/n/optimization.log``::
-
-    Generation1: Best Fitness = 5.864228e+00
-    Generation2: Best Fitness = 5.864228e+00
-    Generation3: Best Fitness = 4.488934e+00
-    Generation4: Best Fitness = 3.793744e+00
-    Generation5: Best Fitness = 3.652047e+00
-    Generation6: Best Fitness = 3.652047e+00
-    Generation7: Best Fitness = 3.652047e+00
-    Generation8: Best Fitness = 3.452999e+00
-    Generation9: Best Fitness = 3.180878e+00
-    Generation10: Best Fitness = 1.392501e+00
-    Generation11: Best Fitness = 1.392501e+00
-    Generation12: Best Fitness = 1.392501e+00
-    Generation13: Best Fitness = 1.392501e+00
-    Generation14: Best Fitness = 7.018051e-01
-    Generation15: Best Fitness = 7.018051e-01
-    Generation16: Best Fitness = 7.018051e-01
-    Generation17: Best Fitness = 7.018051e-01
-    Generation18: Best Fitness = 7.018051e-01
-    Generation19: Best Fitness = 6.862063e-01
-    Generation20: Best Fitness = 6.862063e-01
-
-* If you want to continue from where you stopped in the last parameter search,
-
-.. code-block:: python
-
-    from biomass import optimize_continue
-
-    optimize_continue(
-        model, x_id=1, options={
-            "popsize": 3,
-            "max_generation": 200,
-            "allowable_error": 0.5,
-            "local_search_method": "DE",
-            "maxiter": 50,
-        }
-    )
-
-* If you want to search multiple parameter sets (e.g., from 1 to 10) simultaneously,
+* Set simulation conditions and the corresponding experimental data in ``observable.py``
+* Define an objective function to be minimized (:func:`objective`) in ``fitness.py``
+* Set lower/upper bounds of parameters to be estimated in ``set_search_param.py``
 
 .. code-block:: python
 
     from biomass import optimize
+    
+    optimize(model, x_id=1, optimizer_options={"workers": -1})
 
-    optimize(
-        model, x_id=range(1, 11), options={
-            "popsize": 5,
-            "max_generation": 100,
-            "allowable_error": 0.5,
-            "local_search_method": "DE",
-            "maxiter": 50,
-        }
-    )
+The temporary result will be saved in ``out/_tmp{n}/`` after each iteration.
 
-Using external optimizers
-^^^^^^^^^^^^^^^^^^^^^^^^^
+Progress list: ``out/_tmp{n}/optimization.log``::
 
-You can also use external optimization methods to determine model parameters.
-Below is an example of using ``scipy.optimize.differential_evolution`` for parameter estimation.
+    differential_evolution step 1: f(x)= 4.96181
+    differential_evolution step 2: f(x)= 3.555
+    differential_evolution step 3: f(x)= 2.50626
+    differential_evolution step 4: f(x)= 2.00657
+    differential_evolution step 5: f(x)= 1.83556
+    differential_evolution step 6: f(x)= 1.28031
+    differential_evolution step 7: f(x)= 0.973207
+    differential_evolution step 8: f(x)= 0.741667
+    differential_evolution step 9: f(x)= 0.741667
+    differential_evolution step 10: f(x)= 0.735682
+    differential_evolution step 11: f(x)= 0.717266
+    differential_evolution step 12: f(x)= 0.603178
+    differential_evolution step 13: f(x)= 0.56934
+    differential_evolution step 14: f(x)= 0.56934
+    differential_evolution step 15: f(x)= 0.549331
+    differential_evolution step 16: f(x)= 0.459069
+    differential_evolution step 17: f(x)= 0.447772
+    differential_evolution step 18: f(x)= 0.430385
+    differential_evolution step 19: f(x)= 0.37085
+    differential_evolution step 20: f(x)= 0.37085
 
-.. code-block:: python
-
-    from scipy.optimize import differential_evolution
-
-    from biomass import Model
-    from biomass.estimation import ExternalOptimizer
-    from biomass.models import Nakakuki_Cell_2010
-
-    model = Model(Nakakuki_Cell_2010.__package__).create()
-    optimizer = ExternalOptimizer(model, differential_evolution)
-
-    def obj_fun(x):
-        """Objective function to be minimized."""
-        return optimizer.get_obj_val(x)
-
-    res = optimizer.run(
-        obj_fun,
-        [(0, 1) for _ in range(len(model.problem.bounds))],
-        strategy="best1bin",
-        maxiter=50,
-        tol=1e-4,
-        mutation=0.1,
-        recombination=0.5,
-        disp=True,
-        polish=False,
-        workers=-1,
-    )
-
-* Import the solution of the optimization (`res.x`) and visualize the result.
-
-.. code-block:: python
-
-
-    from biomass import run_simulation
-
-    param_values = model.problem.gene2val(res.x)
-    optimizer.import_solution(param_values, x_id=0)
-    run_simulation(model, viz_type="0")
+.. note::
+    For detailed information about ``optimizer_options``, please refer to `scipy docs <https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.differential_evolution.html>`.
 
 Data export and visualization
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
